+++ src/Doer/TestEditor.php
@@ -14,13 +14,16 @@ class WpTesting_Doer_TestEditor extends WpTesting_Doer_AbstractDoer
         $this->sessionInit(__CLASS__)->wp
             ->enqueuePluginStyle('wpt_admin', 'css/admin.css')
             ->enqueuePluginScript('wpt_test_edit_fix_styles',  'js/test-edit-fix-styles.js',        array('jquery'), false, true)
-            ->enqueuePluginScript('field_selection',           'js/vendor/kof/field-selection.js',  array('jquery'), false, true)
-            ->enqueuePluginScript('wpt_test_edit_formulas',    'js/test-edit-formulas.js',          array('jquery'), false, true)
+            ->enqueuePluginScript('field_selection',           'js/vendor/kof/field-selection.js',  array(), false, true)
+            ->enqueuePluginScript('wpt_test_edit_formulas',    'js/test-edit-formulas.js',          array('jquery', 'field_selection'), false, true)
+            ->enqueuePluginScript('json3',                     'js/vendor/bestiejs/json3.min.js',   array(), false, true)
+            ->enqueuePluginScript('wpt_test_quick_scores',     'js/test-quick-scores.js',           array('jquery', 'lodash'), false, true)
+            ->enqueuePluginScript('wpt_test_quick_questions',  'js/test-quick-questions.js',        array('jquery', 'json3'), false, true)
             ->addAction('post_submitbox_misc_actions', array($this, 'renderSubmitMiscActions'))
             ->addAction('media_buttons',               array($this, 'renderContentEditorButtons'))
-            ->addMetaBox('wpt_edit_questions', 'Edit Questions',    array($this, 'renderEditQuestions'), 'wpt_test')
-            ->addMetaBox('wpt_add_questions',  'Add New Questions', array($this, 'renderAddQuestions'),  'wpt_test')
-            ->addMetaBox('wpt_edit_formulas',  'Edit Formulas',     array($this, 'renderEditFormulas'),  'wpt_test')
+            ->addMetaBox('wpt_edit_questions', __('Edit Questions and Scores', 'wp-testing'),    array($this, 'renderEditQuestions'), 'wpt_test')
+            ->addMetaBox('wpt_add_questions',  __('Add New Questions', 'wp-testing'), array($this, 'renderAddQuestions'),  'wpt_test')
+            ->addMetaBox('wpt_edit_formulas',  __('Edit Formulas', 'wp-testing'),     array($this, 'renderEditFormulas'),  'wpt_test')
             ->addAction('admin_notices', array($this, 'printAdminMessages'))
             ->addAction('save_post',     array($this, 'saveTest'), 10, 2)
         ;
@@ -54,10 +57,12 @@ class WpTesting_Doer_TestEditor extends WpTesting_Doer_AbstractDoer
         $test = new WpTesting_Model_Test($item);
         $this->output('Test/Editor/edit-questions', array(
             'scales'      => $test->buildScalesWithRange(),
+            'answers'     => $test->buildAnswers(),
             'questions'   => $test->buildQuestions(),
             'isWarnOfSettings'   => $test->isWarnOfSettings(),
             'memoryWarnSettings' => $test->getMemoryWarnSettings(),
             'isUnderApache'      => $this->isUnderApache(),
+            'canEditScores'      => $test->canEditScores(),
         ));
     }
 
@@ -70,7 +75,6 @@ class WpTesting_Doer_TestEditor extends WpTesting_Doer_AbstractDoer
         $this->output('Test/Editor/add-questions', array(
             'addNewCount' => WpTesting_Model_Question::ADD_NEW_COUNT,
             'startFrom'   => $test->buildQuestions()->count(),
-            'scales'      => $test->buildScales(),
         ));
     }
 
@@ -114,7 +118,7 @@ class WpTesting_Doer_TestEditor extends WpTesting_Doer_AbstractDoer
             $problems = $test->validate();
             $test->store(true);
         } catch (fValidationException $e) {
-            $title = 'Test data not saved';
+            $title = __('Test data not saved', 'wp-testing');
             $this->wp->dieMessage(
                 $this->render('Test/Editor/admin-message', array(
                     'title'   => $title,