+++ src/Template/Test/Editor/add-questions.php
@@ -1,18 +1,23 @@
-<table class="widefat wpt_questions">
-    <tr>
-        <th></th>
-    <?php foreach($scales as $i => $scale): /* @var $scale WpTesting_Model_Scale */ ?>
-        <th class="wpt_scale <?php echo ($i%2) ? '' : 'alternate' ?>">
-            <?php echo $scale->getTitle() ?>
-        </th>
-    <?php endforeach ?>
-    </tr>
+<div id="wpt_quick_fill_questions" class="hide-if-no-js wp-hidden-children">
+
+<h4>
+    <a href="#wpt-quick-fill-questions" class="toggle"><?php echo __('Quick Fill From Text', 'wp-testing') ?></a>
+</h4>
+
+<div class="wp-hidden-child">
+<textarea rows="20" cols="50" placeholder="<?php echo __('Paste here your questions and they will fill fields below. Numbers and other indexes will be stripped automatically.', 'wp-testing') ?>"></textarea>
+<input type="button" class="button" value="<?php echo __('Quick Fill From Text', 'wp-testing') ?>" />
+</div>
+
+</div>
+
+<table class="widefat wpt_questions" data-start-from="<?php echo $startFrom ?>">
 <?php foreach(range(0, $addNewCount) as $i): ?>
     <tr class="wpt_question <?php echo ($i%2) ? 'alternate' :'bar' ?>">
         <th class="wpt_number">
             *
         </th>
-        <td class="wpt_title" colspan="<?php echo count($scales) ?>">
+        <td class="wpt_title">
             <input type="text" name='wpt_question_title[<?php echo json_encode(array(
                 'i'  => $startFrom + $i,
                 'id' => '',