+++ src/Template/Test/Editor/content-editor-buttons.php
@@ -1,12 +1,12 @@
 <a href="#wpt_add_questions" class="button wpt-button add">
     <span class="wp-media-buttons-icon dashicons dashicons-plus"></span>
-    <?php echo 'Add New Questions' ?>
+    <?php echo __('Add New Questions', 'wp-testing') ?>
 </a>
 <a href="#wpt_edit_questions" class="button wpt-button edit">
     <span class="wp-media-buttons-icon dashicons dashicons-edit"></span>
-    <?php echo 'Edit Questions' ?>
+    <?php echo __('Edit Questions and Scores', 'wp-testing') ?>
 </a>
 <a href="#wpt_edit_formulas"  class="button wpt-button edit">
     <span class="wp-media-buttons-icon dashicons dashicons-edit"></span>
-    <?php echo 'Edit Formulas' ?>
+    <?php echo __('Edit Formulas', 'wp-testing') ?>
 </a>