+++ wp-testing.php
@@ -3,10 +3,12 @@
  * Plugin Name: Wp-testing
  * Plugin URI: http://wordpress.org/extend/plugins/wp-testing/
  * Description: Helps to create psychological tests.
- * Version: 0.4.2
+ * Version: 0.5
  * Author: Alexander Ustimenko
  * Author URI: http://ustimen.co
  * License: GPL3
+ * Text Domain: wp-testing
+ * Domain Path: /languages
  */
 
 require_once dirname(__FILE__) . '/src/WordPressFacade.php';